Refactor shipPlacement and related methods
To prepare for rotation functionality, the shipPlacement function has
been refacotred. An array of coordinates are generated for the ship
when the player attempts to place it. Depending on the orientation of
the ship, these coordinates will reflect the would-be position of the
ship. This has replaced the placeholder "y+1" method in previous
versions.

The orientation method has been tested, but the tests related to placing
the ship will require updating.