Make AI more "intelligent"
When the AI hits a ship, it generates a stack of moves corresponding to
coordinates that are adjacent to the spot of the hit. On each
successive hit, the move that lead to the hit is re-added to the
stack until it leads to a miss. The stack is cleared when the attacked
ship is sunk.